Move Pagination component to components folder
