+++ src/components/TabLists/PaginationComponent/index.js
@@ -1,0 +1,3 @@
+import PaginationComponent from './PaginationComponent';
+
+export default PaginationComponent;
