+++ src/components/TabLists/PaginationComponent/styles.js
@@ -1,0 +1,14 @@
+
+// Styles
+export default theme => ({
+  button: {
+    padding: `${theme.spacing.unit / 2}px ${theme.spacing.unit}px`,
+    margin: theme.spacing.unit / 2,
+    minWidth: 'auto',
+  },
+  pageItem: {
+    margin: 0,
+    marginLeft: theme.spacing.unit / 2,
+    marginRight: theme.spacing.unit / 2,
+  },
+});
